+++ packages/manager-interface/src/legacy/components/pages/Overview.jsx
@@ -1,0 +1,14 @@
+import React, { Fragment } from 'react';
+import GetStarted from '../../containers/GetStarted';
+import Ranking from '../../containers/Ranking';
+
+const Overview = () => {
+  return (
+    <Fragment>
+      <GetStarted />
+      <Ranking />
+    </Fragment>
+  );
+};
+
+export default Overview;
